refactor: extract VehicleForm component from Garage.tsx
Extracts the vehicleSchema, VehicleFormData type, and the entire add/edit
vehicle form JSX block from `src/components/Garage.tsx` into a separate
`src/components/VehicleForm.tsx` file.

This improves maintainability and readability by reducing the size and
complexity of the Garage component. Garage.tsx now simply manages the
global vehicle state and conditionally renders the new VehicleForm
component, passing down the necessary state as props.

Tested with `npm run lint` and `npm run build` to ensure no type issues
or compilation errors were introduced. Verified frontend functionality
via Playwright screenshot.
